Fix issues related to .pth rewrite addition
Fix file discovery to actually recurse the entire virtualenv tree.
Originally, the call to extend the list of directories with
subdirectories was missing. The result was that only top the first two
levels of the directory structure were actually processed. The test is
updated to ensure the full file tree is processed.

The stack of directories was also being mismanaged in a way that could
result in pop from an empty list. Notably, the search would fail when
processing the last batch of files. This was due to the loop condition
which engaged the loop body if there were directory _or_ files to be
processed. If there were only files and all directories had been
consumed then the last iteration would error when trying to pop from the
empty directory list.