Change activate script regexes to fix quote chars
A patch introduced in virtualenv 20.0.0 changed double quotes to single
quotes for the activate script variables that contain the embedded path
of the virtualenv. The existing tests were able to detect the regression
when run with the latest virtualenv. This patch modifies the regexes
used to detect the embedded path by checking for both kinds of quote
characters. Additionally, the output format of modified paths now uses
single quotes anywhere that virtualenv does so the outputs match.