Add provisional stdlib venv support
This patch modifies the way the bash activate script is handled to work
with both virtualenv and the python3 bundled venv. The venv version
had a recent change for windows support that caused it to diverge from
how virtualenv solved the same cygwin related problems. So there are now
two totally different bash activation scripts depending on which tool
you use.

The venv style puts the literal path text in several location whereas
every other script in both venv and virtualenv puts the literal text
only once in a variable and then references the variable. To change
multiple path strings I had to rewrite the path replacement mechanism
for bash to bulk rewrite every line that contains the path rather than a
targeted line replacement like before.

I'm not sure how much more venv will diverge from virtualenv in the
future but this should enable it to work for now.