Fix relocation for virtualenv>=20.26.6
Virtualenv versions 20.26.6 and greater include a change to the activate
script templates. Previously, the full path to the virtualenv that is
rendered into the activation script was quoted. The newer versions
render unquoted paths. The regular expressions used to match these paths
now conditionally matches the quotes which should work for both old and
new activation scripts.