Fix handling of venv bash activation scripts
The regular expression used to capture the venv's path in a bash
activation script was capturing too much. This caused the rewrite that
happens during relocate to leave an instance of original path but only
in a branch of bash code that is activated when running on Windows.

This patch fixes a bug that was actually present in all regular
expressions used to find paths in activation scripts. The bug was
previously benign because all virtualenv activation scripts and all but
the bash script for venv contain exactly one instance of the matched
path which was correctly changed during relocation. Now all scripts use
an improved pattern and will automatically handle multiple path
instances if they appear over time.